requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.1
Babel==2.8.0
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-phonenumber-field==4.0.0
django-restframework==0.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumberslite==8.12.1
psycopg2==2.8.5
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
rope==0.16.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.11.2
